backend/requirements.txt
fastmcp>=0.2.0
fastapi>=0.115.0
uvicorn>=0.30.0
tree-sitter>=0.21.0
tree-sitter-python>=0.21.0
tree-sitter-javascript>=0.21.0
ruff>=0.6.0
mypy>=1.11.0
rope>=1.13.0
docker>=7.1.0
groq>=0.11.0
python-dotenv>=1.0.0
pydantic>=2.9.0
sqlalchemy>=2.0.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
